Fix admin_panel . Add Search class
Solved admin_pane always going into else statement

+++ admin_panel.php
@@ -14,8 +14,9 @@
 <body>
     <form action="" method="get">
         <label for="date">Date</label>
-        <input type="date" name="date" id="date">
-        <input type="submit" value="show">
+        <input type="date" name="date" id="date" value="<?php echo date('Y-m-d', strtotime('-1 day')); ?>">
+
+        <input type="submit" value="Arata">
     </form>
 </body>
 </html>
@@ -26,24 +27,26 @@ require_once 'classes/Crud.php';
 
 $show_entries = new Crud;
 
-$date = $_GET['date'];
-
-echo $date;
 
 // $show_reg = $show_entries->get_numbers();
+if(!empty($_GET['date'])){
 
-$show_reg = $show_entries->get_numbers($_GET['date']);
+    $show_reg = $show_entries->get_numbers($_GET['date']);
 
-foreach ($show_reg as $row) {
-    echo "<form action=delete_id.php method='GET'";
-    echo '<p style="display: hidden"'.$row['id'] . '</p>';
-    echo '<h4>'.$row['call_nr'].'</h4>';
-    echo '<p>'.$row['date'].'</p>';
-    echo '<p>'.$row['name'].'</p>';
-    echo '<p>'.$row['truck_number'].'</p>';
-    echo '<p>'.$row['phone_number'].'</p>';
-    echo "<a href='delete_id.php?id=".$row['id']."'>Delete</a> <br>";
-    echo '<hr>';
-}
+    if(!empty($show_reg)){
+
+        foreach ($show_reg as $row) {
+            echo "<form action=delete_id.php method='GET'";
+            echo '<p style="display: hidden"'.$row['id'] . '</p>';
+            echo '<h4>'.$row['call_nr'].'</h4>';
+            echo '<p>'.$row['date'].'</p>';
+            echo '<p>'.$row['name'].'</p>';
+            echo '<p>'.$row['truck_number'].'</p>';
+            echo '<p>'.$row['phone_number'].'</p>';
+            echo "<a href='delete_id.php?id=".$row['id']."'>Delete</a> <br>";
+            echo '<hr>';
+        } 
+        } else  echo 'Nu exista inregistrari cu o data egala sau mai mare decat data selectata';
+    } 
 
 ?>
